+++ test/e2e/tests/account/migrate-old-vault.spec.ts
@@ -3,7 +3,7 @@ import { defaultGanacheOptions, withFixtures } from '../../helpers';
 import FixtureBuilder from '../../fixture-builder';
 import { Driver } from '../../webdriver/driver';
 import { Ganache } from '../../seeder/ganache';
-import HomePage from '../../page-objects/pages/homepage';
+import HomePage from '../../page-objects/pages/home/homepage';
 import { loginWithBalanceValidation } from '../../page-objects/flows/login.flow';
 
 describe('Migrate vault with old encryption', function (this: Suite) {
