+++ test/e2e/tests/confirmations/transactions/native-send-redesign.spec.ts
@@ -7,7 +7,7 @@ import {
   WINDOW_TITLES,
 } from '../../../helpers';
 import TokenTransferTransactionConfirmation from '../../../page-objects/pages/confirmations/redesign/token-transfer-confirmation';
-import HomePage from '../../../page-objects/pages/homepage';
+import HomePage from '../../../page-objects/pages/home/homepage';
 import SendTokenPage from '../../../page-objects/pages/send/send-token-page';
 import TestDapp from '../../../page-objects/pages/test-dapp';
 import { Driver } from '../../../webdriver/driver';
