+++ test/e2e/tests/privacy/account-tracker-api-usage.spec.ts
@@ -8,7 +8,7 @@ import {
   withFixtures,
 } from '../../helpers';
 import { Mockttp } from '../../mock-e2e';
-import HomePage from '../../page-objects/pages/homepage';
+import HomePage from '../../page-objects/pages/home/homepage';
 import { loginWithoutBalanceValidation } from '../../page-objects/flows/login.flow';
 
 async function mockInfura(mockServer: Mockttp): Promise<MockedEndpoint[]> {
