+++ test/e2e/tests/privacy/polling.spec.ts
@@ -5,7 +5,7 @@ import { expect } from '@playwright/test';
 import FixtureBuilder from '../../fixture-builder';
 import { defaultGanacheOptions, withFixtures } from '../../helpers';
 import { Mockttp } from '../../mock-e2e';
-import HomePage from '../../page-objects/pages/homepage';
+import HomePage from '../../page-objects/pages/home/homepage';
 import { loginWithoutBalanceValidation } from '../../page-objects/flows/login.flow';
 
 const infuraMainnetUrl =
