+++ test/e2e/tests/tokens/add-token-using-search.ts
@@ -2,7 +2,7 @@ import { MockedEndpoint, Mockttp } from 'mockttp';
 import { defaultGanacheOptions, withFixtures } from '../../helpers';
 import FixtureBuilder from '../../fixture-builder';
 import { CHAIN_IDS } from '../../../../shared/constants/network';
-import HomePage from '../../page-objects/pages/homepage';
+import HomePage from '../../page-objects/pages/home/homepage';
 import { loginWithoutBalanceValidation } from '../../page-objects/flows/login.flow';
 
 describe('Add existing token using search', function () {
