+++ test/e2e/tests/transaction/stuck-approved-transaction.spec.ts
@@ -4,7 +4,8 @@ import FixtureBuilder from '../../fixture-builder';
 import { Driver } from '../../webdriver/driver';
 import { Ganache } from '../../seeder/ganache';
 import { loginWithBalanceValidation } from '../../page-objects/flows/login.flow';
-import HomePage from '../../page-objects/pages/homepage';
+import ActivityListPage from '../../page-objects/pages/home/activity-list';
+import HomePage from '../../page-objects/pages/home/homepage';
 
 describe('Editing Confirm Transaction', function (this: Suite) {
   it('approves a transaction stuck in approved state on boot', async function () {
@@ -25,10 +26,10 @@ describe('Editing Confirm Transaction', function (this: Suite) {
       }) => {
         await loginWithBalanceValidation(driver, ganacheServer);
 
-        const homePage = new HomePage(driver);
-        await homePage.goToActivityList();
-        await homePage.check_completedTxNumberDisplayedInActivity();
-        await homePage.check_txAmountInActivity();
+        new HomePage(driver).goToActivityList();
+        const activityList = new ActivityListPage(driver);
+        await activityList.check_completedTxNumberDisplayedInActivity();
+        await activityList.check_txAmountInActivity();
       },
     );
   });
